refactor: split controller tests into a sibling test file
The fence-neutralization rounds pushed agent/controller.rs past the R8
1400-LOC hard cap; same directory-module split as tools_quality and
evidence (mod.rs 694 + test.rs 753, 25 tests before and after, pure
movement).
